feat(application): 增加浏览历史记录更新文章阅读量功能- 在 ViewHistoryCreatedDomainEventSubscriber 中添加更新文章阅读量的逻辑 - 引入 UpdateArticleViewCountCmd 和 Mediator 用于发送更新命令 - 使用 Optional 包装更新命令的构建过程，确保安全性
