+++ only4-KSP-cap4j-domain/src/main/java/com/only4/domain/aggregates/role/events/RoleInfoUpdatedDomainEvent.java
@@ -9,14 +9,14 @@ import org.netcorepal.cap4j.ddd.domain.aggregate.annotation.Aggregate;
 import org.netcorepal.cap4j.ddd.domain.event.annotation.DomainEvent;
 
 /**
- * Role.RoleInfoUpdatedDomainEvent领域事件
+ * Role.UpdatedRoleInfoDomainEvent领域事件
  * 角色信息已更新
  *
  * @author cap4j-ddd-codegen
- * @date 2025/02/20
+ * @date 2024/12/15
  */
 @DomainEvent(persist = false)
-@Aggregate(aggregate = "Role", name = "RoleInfoUpdatedDomainEvent", type = Aggregate.TYPE_DOMAIN_EVENT, description = "")
+@Aggregate(aggregate = "Role", name = "UpdatedRoleInfoDomainEvent", type = Aggregate.TYPE_DOMAIN_EVENT, description = "")
 @Data
 @Builder
 @AllArgsConstructor
