--- only4-KSP-cap4j-domain/src/main/java/com/only4/domain/aggregates/role/events/UpdatedRoleInfoDomainEvent.java
@@ -1,26 +1,0 @@
-package com.only4.domain.aggregates.role.events;
-
-import com.only4.domain.aggregates.role.Role;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.netcorepal.cap4j.ddd.domain.aggregate.annotation.Aggregate;
-import org.netcorepal.cap4j.ddd.domain.event.annotation.DomainEvent;
-
-/**
- * Role.UpdatedRoleInfoDomainEvent领域事件
- * todo: 领域事件说明
- *
- * @author cap4j-ddd-codegen
- * @date 2024/12/15
- */
-@DomainEvent(persist = false)
-@Aggregate(aggregate = "Role", name = "UpdatedRoleInfoDomainEvent", type = Aggregate.TYPE_DOMAIN_EVENT, description = "")
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class UpdatedRoleInfoDomainEvent {
-    Role role;
-}
